+++ CHANGES.md
@@ -11,16 +11,15 @@ Github. Parentheses after an item show the name or github id of the contributor
 of that change.
 
 
-
-
 ## 1.0.16.dev0 (Work In Progress)
 
 ### New:
 
+- `ImageDataBunch.single_from_classes` to allow single image predictions
+
 ### Changed:
 
 ### Fixed:
-
 
 
 ## 1.0.15 (2018-10-28)